requirements.txt
streamlit
pandas
python-dotenv
google-generativeai
PyPDF2
matplotlib